refactor: Move custom file picker toggle into Files & storage section
