+++ app/src/main/res/values/strings.xml
@@ -672,10 +672,8 @@ The image dimensions must be as follows:
     <string name="settings_system_export_debug_logs_export_failed">Failed to export logs</string>
     <string name="settings_system_export_debug_logs_export_success">Logs exported</string>
 
-    <!-- Settings - System Tab - Files Section -->
-    <string name="settings_system_files">Files</string>
-    <string name="settings_system_custom_file_picker">Custom file picker</string>
-    <string name="settings_system_custom_file_picker_description">Browse storage with Morphe\'s built-in file picker instead of the system default</string>
+    <!-- Settings - System Tab - Files & Storage Section -->
+
 
     <!-- File Picker -->
     <string name="file_picker_title">Select file</string>
@@ -691,8 +689,8 @@ The image dimensions must be as follows:
     <string name="file_picker_sort_date_desc">Date (newest first)</string>
     <string name="file_picker_sort_date_asc">Date (oldest first)</string>
 
-    <!-- Settings - System Tab - Storage Management Section -->
-    <string name="settings_system_storage_management">Storage management</string>
+    <!-- Settings - System Tab - Files & Storage Section -->
+    <string name="settings_system_files">Files &amp; storage</string>
     <string name="settings_system_apks_size">Total size: %s</string>
     <string name="settings_system_apk_item_info" translatable="false">v%s • %s</string>
 
@@ -734,6 +732,10 @@ The image dimensions must be as follows:
     <string name="settings_system_selected_patches_title">Selected patches (%s)</string>
     <string name="settings_system_patch_options_title">Patch options (%s)</string>
     <string name="settings_system_no_patches_or_options">No patches or options saved</string>
+
+    <!-- Settings - System Tab - Custom file picker -->
+    <string name="settings_system_custom_file_picker">Custom file picker</string>
+    <string name="settings_system_custom_file_picker_description">Browse storage with Morphe\'s built-in file picker instead of the system default</string>
 
     <!-- Settings - System Tab - About Section -->
     <string name="settings_system_about">About</string>
